refactor(kafka): split kafka.go into write, read, reader, and config files
Keep the task lifecycle in kafka.go and move producer, consumer poll, consumer wrappers, and librdkafka config builders so each concern stays in a smaller file.
